feat(reservations): Support for a 'v4' reservation report (#2589)
The V4 reservation report allows an external source (ie. file in s3) to
define how surplus regional reservations should be shared with their
AZ equivalents.

This is likely to remain a Netflix-specific variant.

+++ clouddriver-aws/src/main/groovy/com/netflix/spinnaker/clouddriver/aws/model/AmazonReservationReport.groovy
@@ -218,16 +218,16 @@ class AmazonReservationReport implements ReservationReport {
     int index
 
     @JsonView(Views.V3.class)
-    int totalRegionalSurplusForFamily
+    Integer totalRegionalSurplusForFamily
 
     @JsonView(Views.V3.class)
-    int totalShortfallForFamily
+    Integer totalShortfallForFamily
 
     @JsonView(Views.V3.class)
-    double percentageOfShortfall
+    Double percentageOfShortfall
 
     @JsonView(Views.V3.class)
-    int portionOfAvailableSurplus
+    Double portionOfAvailableSurplus
   }
 
   static class DescendingOverallReservationDetailComparator implements Comparator<OverallReservationDetail> {
